requirements.txt
paho-mqtt
pyyaml
flask